* Attempt to removed warnings in Windows X64
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8421 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/atermpp/include/mcrl2/atermpp/aterm_appl.h
@@ -208,7 +208,7 @@ namespace atermpp
       /// \param t A term
       /// \param i A positive integer
       /// \return The term with one of its arguments replaced.
-      term_appl<Term> set_argument(Term t, unsigned int i)
+      term_appl<Term> set_argument(Term t, size_t i)
       {
         return ATsetArgument(appl(), t, i);
       }
@@ -217,7 +217,7 @@ namespace atermpp
       /// \deprecated
       /// \param i A positive integer
       /// \return The argument with the given index.
-      aterm argument(unsigned int i) const
+      aterm argument(size_t i) const
       {
         return aterm(ATgetArgument(appl(), i));
       }
